CORDA-3972: Revert API changes
